update eeprom stuff

+++ inc/soc.h
@@ -1,5 +1,5 @@
 #ifndef _SOC_H_
 #define _SOC_H_
 void SOC_Init(void);
-uint16_t SOC_Estimate(void);
+uint32_t SOC_Estimate(void);
 #endif
